chore(test): setup unit testing
- update es5 to es6
- implementation of the arrow function

[Finishes #165413095]
