+++ .gitignore
@@ -34,3 +34,8 @@ node_modules
 # Optional REPL history
 .node_repl_history
 
+#Report directory
+.nyc_output
+
+package-lock.json
+
