+++ app.js
@@ -1,0 +1,85 @@
+import express from 'express';
+import bodyParser from 'express';
+import session from 'express-session';
+import cors from 'cors';
+import errorhandler from 'errorhandler';
+import mongoose from 'mongoose';
+
+const isProduction = process.env.NODE_ENV === "production";
+
+// Create global app object
+const app = express();
+
+app.use(cors());
+
+// Normal express config defaults
+app.use(require("morgan")("dev"));
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(bodyParser.json());
+
+app.use(require("method-override")());
+app.use(express.static(__dirname + "/public"));
+
+app.use(
+    session({
+        secret: "authorshaven",
+        cookie: { maxAge: 60000 },
+        resave: false,
+        saveUninitialized: false
+    })
+);
+
+if (!isProduction) {
+    app.use(errorhandler());
+}
+
+if (isProduction) {
+    mongoose.connect(process.env.MONGODB_URI);
+} else {
+    mongoose.connect("mongodb://localhost/conduit");
+    mongoose.set("debug", true);
+}
+
+require("./models/User");
+
+app.use(require("./routes"));
+
+/// catch 404 and forward to error handler
+app.use(function (req, res, next) {
+    const err = new Error("Not Found");
+    err.status = 404;
+    next(err);
+});
+
+/// error handlers
+
+// development error handler
+// will print stacktrace
+if (!isProduction) {
+    app.use(function (err, req, res, next) {
+        console.log(err.stack);
+
+        res.status(err.status || 500);
+
+        res.json({
+            errors: {
+                message: err.message,
+                error: err
+            }
+        });
+    });
+}
+
+// production error handler
+// no stacktraces leaked to user
+app.use(function (err, req, res, next) {
+    res.status(err.status || 500);
+    res.json({
+        errors: {
+            message: err.message,
+            error: {}
+        }
+    });
+});
+
+export default app;
